fix: type mismatch in infman readDisk for cross-platform
FreeBSD has int64 Bavail while Linux has uint64. Cast to uint64
to ensure cross-platform compilation works.

+++ src/infman/sysinfo.go
@@ -265,7 +265,7 @@ func readDisk(path string) (uint64, uint64, bool) {
 		return 0, 0, false
 	}
 	total := stat.Blocks * uint64(stat.Bsize)
-	avail := stat.Bavail * uint64(stat.Bsize)
+	avail := uint64(stat.Bavail) * uint64(stat.Bsize)
 	return total - avail, total, true
 }
